build: move openapi/patterns under /api
The goal is to remove dependency on the common/ package from /api
