+++ api/openapi/patterns/anchored.go
