+++ api/openapi/patterns/distribution.go
