+++ api/openapi/patterns/dns1123.go
