+++ api/openapi/patterns/generic.go
