+++ api/openapi/patterns/net.go
