+++ api/v1alpha1/clusterconfig_types.go
@@ -11,7 +11,7 @@ import (
 	"k8s.io/utils/ptr"
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
 
-	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/openapi/patterns"
+	"github.com/d2iq-labs/capi-runtime-extensions/api/openapi/patterns"
 )
 
 const (
